refactor(code_gen): extract execute() into stmt.rs
